feat(components): split Dropdown menu up
